==> data/ref/docs/03-sub.qmd <==
## preamble

### setup

```{r load-libs, message=FALSE, warning=FALSE}
library(dplyr)
library(tidyr)
library(irlba)
library(scran)
library(scater)
library(igraph)
library(AUCell)
library(harmony)
library(ggplot2)
library(tidytext)
library(patchwork)
library(BiocParallel)
bp <- MulticoreParam(th <- 50)
source("utils.R")
set.seed(250309)
```

### loading

```{r load-data}
sce <- readRDS("../outs/sce-lv1.rds")
ids <- list(
    epi="epi",
    str="fib|EC|SMC",
    imm="BC|PC|TC|mye|mast")
idx <- lapply(ids, grep, sce$lv1)
lys <- lapply(idx, \(.) sce[, .])
names(sub) <- sub <- names(lys)
sapply(lys, \(.) table(.$sid))
sapply(lys, ncol)
```

```{r echo=FALSE}
stopifnot(sum(sapply(lys, ncol)) == sum(!is.na(sce$lv1)))
```

## selection

```{r sel}
sel <- lapply(lys, \(sce) {
    tbl <- modelGeneVar(sce, block=sce$pid, BPPARAM=bp)
    tbl <- tbl[-grep("^H[0-9]|^RP(L|S)|^MT-", rownames(tbl)), ]
    sel <- lapply(tbl$per.block, getTopHVGs, n=2e3)
    sel <- sapply(sel, \(.) rownames(sce) %in% .)
    rownames(sce)[rowSums(sel) > 1]
})
sapply(sel, length)
```

## integration

```{r int}
lys <- mapply(sce=lys, sel=sel, \(sce, sel) {
    reducedDim(sce, "PCA") <- RunHarmony(
        meta_data=sce$run, verbose=FALSE, ncores=th,
        prcomp_irlba(t(logcounts(sce[sel, ])), 30)$x)
    runUMAP(sce, dimred="PCA", BPPARAM=bp)
})
```

## clustering

```{r clu}
.clu <- \(sce, res) {
    g <- buildSNNGraph(sce, use.dimred="PCA", type="jaccard", k=20, BPPARAM=bp)
    k <- cluster_leiden(g, objective_function="modularity", resolution=res)
    l <- letters[seq_along(unique(k <- k$membership))]
    sce$kid <- factor(k, labels=l); sce
}
set.seed(7); lys$epi <- .clu(lys$epi, 1.8)
set.seed(7); lys$str <- .clu(lys$str, 1.4)
set.seed(7); lys$imm <- .clu(lys$imm, 1.0)
```

```{r clu-ns}
# tabulate cell/cluster counts
lapply(lys, \(.) table(.$kid))
sapply(lys, \(.) nlevels(.$kid))
```

## plotting

### compositions

::: {.panel-tabset}

```{r plt-fq-one, results="asis", fig.width=4, fig.height=2}
for (. in names(lys)) {
    p <- .plt_fq(lys[[.]], x="kid", y="sid", id=., pal=.pal)
    cat("#### ", ., "\n"); print(p); cat("\n\n")
}
```

:::

### regression

::: {.panel-tabset}

```{r pcr-kid, results="asis", fig.width=6, fig.height=3}
ids <- c("sum", "run", "sid", "kid")
pal <- c("magenta", "gold", "cyan", "blue")
pcr <- lapply(lys, .pcr, ids)
for (. in sub) {
    p <- .plt_pcr(pcr[[.]]) + scale_color_manual(values=pal)
    cat("#### ", ., "\n"); print(p); cat("\n\n")
}
```

:::

### reduction

::: {.panel-tabset}

```{r plt-map-one, results="asis", fig.width=15, fig.height=5}
ids <- c("run", "sid", "kid")
for (. in sub) {
    p <- .plt_dr(lys[[.]], ids) & ggtitle(.lab(., ncol(lys[[.]])))
    cat("#### ", ., "\n"); print(p); cat("\n\n")
}
```

:::

### contingency

```{r tbl, fig.width=12, fig.height=2.5}
lapply(sub, \(.) {
    ns <- table(lys[[.]]$lv1, lys[[.]]$kid)
    fq <- prop.table(ns, 2)
    invisible({
        hm <- pheatmap::pheatmap(fq, silent=TRUE,
            scale="none", angle_col=0, fontsize=9,
            cellwidth=10, cellheight=10, main=.,
            treeheight_row=5, treeheight_col=5)
        ggplotify::as.ggplot(hm)
    })
}) |> wrap_plots(nrow=1)
```

### markers

```{r mgs-de}
mgs <- lapply(lys, \(sub) findMarkers(
    sub, groups=sub$kid, block=sub$pid, 
    direction="up", add.summary=TRUE, BPPARAM=bp))
```

```{r mgs-top}
# top-20 markers
lapply(sub, \(.) sapply(mgs[[.]], \(df) head(rownames(df)[order(-df$summary.logFC)], 20)))
```

::: {.panel-tabset}

```{r mgs-hm, results="asis", fig.width=15, fig.height=4}
ps <- lapply(sub, \(.) {
    top <- unique(unlist(lapply(mgs[[.]], \(df) rownames(df)[df$Top <= 3])))
    if (length(top) > 150) top <- sample(top, 150)
    mtx <- logcounts(lys[[.]])[top, ]
    .plt_hm(lys[[.]], mtx, by="kid", id=.) + 
        coord_flip(expand=FALSE) + theme(
            axis.text.y=element_text(size=12), 
            axis.text.x=element_text(size=8, angle=90, hjust=1, vjust=0.5))
})
for (. in sub) { cat("#### ", ., "\n"); print(ps[[.]]); cat("\n\n") }
```

:::

```{r}
gs <- list(
    imm=list(
        Bm=c("BCL7A", "CR2", "CD19", "CD40", "CD69", "CD79A", "CD83", "CIITA", "SELL", "IGHD"),
        Bn=c("BANK1", "MS4A1"),
        Ba=c("BIRC5"),
        PC_IgA=c("MZB1", "XBP1", "IGHA1"),
        PC_IgG=c("IGHG1"), 
        PC_IgM=c("IGHM"),
        Th=c("CD2", "CD3D", "CD4", "ICOS", "CCR4"),
        Tc=c("CD8A", "GZMA", "GZMK", "NKG7"),
        Tp=c("MKI67", "TOP2A"),
        mast=c("KIT", "IL1RL1"),
        cDC2_CCR7=c("CCR7"),
        cDC1=c("LAMP3", "CLEC10A", "ITGAX"),
        nphil=c("IL1B", "S100A8", "S100A9", "MMP9"),
        TRM=c("CD14", "CD68", "CD163", "APOE", "APOC1", "CTSD", "C1QA", "C1QB", "C1QC")
    ),
    epi=list(
        epi.EE=c(
            c("BEST4", "CA7", "OTOP2", "CTFR", "CA1", "GPT", "MUC17", "TFF1",
        "CD55", "TM4SF1", "DUOX2", "DUOXA1", "APOA1",
        "CHGA", "PCSK1N", "SCT", "SCGN", "NEUROD1", "SOX4",  
        "FCGBP", "ZG16", "SLC26A3", "AQP8", "CEACAM7", "LGR5",
        "RGMB", "ASCL2", "OLFM4", "PCLAF", "PCNA", 
        "SH2D6", "LRMP", "MATK", "FYB1", "HPGDS", "MUC6", "PGC", "AQP5", "BPIFB1"),
            "CHGA", "CHGB", "ENPP2", "SST", "PYY"),
        epi.goblet=c("TFF3", "MUC2", "MUC5AC"),
        epi.entero=c("FABP1", "FABP2", "ALDOB", "SLC26A2", "CA1", "CA4", "CA7", "KRT19", "KRT20"),
        epi.stress=c("DUOX2", "DUOXA2", "CTSE", "KLK10"),
        `epi.goblet-progen`=c("REG4"),
        epi.stem=c("OLFM4", "KCNQ1"),
        epi.prolif=c("MKI67", "TOP2A", "DNMT1"),
        epi.stem_LGR5.1=c("NOTUM", "AXIN2", "ZBTB4", "MCTP1", "ASCL2", "LGR5"),
        epi.stem_LGR5.2=c("CLDN2", "ODAM", "CCL15", "GHR"),
        epi.4=c("TMEM63C"),
        `epi.paneth-like`=c("REG1A", "REG1B", "REG3A", "DEFA5", "DEFA6"),
        epi.1=c("MMP7", "KRT23", "KRT17", "KRT80"),
        epi.6=c("DMXL2", "TUSC3"),
        epi.2=c("TSLP", "ESPN", "ALPI", "ATP12A"),
        epi.3=c("C3", "CISH"),
        epi.5=c("SPARC", "CALD1")
    ),
    str=list(
        SMC=c("MYL9", "MYLK", "MYH11", "ACTA2", "TAGLN"),
        BEC=c("CDH5", "FLT1", "CLDN5"),
        LEC=c("PECAM1", "VWF", "CD34", "CD93", "EGFL7", "KLF2", "LYVE1", "PROX1", "CCL21"),
        fib=c("PTGIS", "SFRP1", "SFRP2", "PI16"),
        fib_TLS=c("C3", "CLU", "LTA3", "CCL19", "CXCL13", "SPIB"),
        IAF_MMP=c("MMP1", "MMP3", "MMP10", "MMP19", "CXCL5", "CXCL8", "SMOX", "ASRGL1"),
        fib_ADAMDEC1=c("ADAMDEC1", "ADAM28", "BMP5", "FGFR4", "CCL11"),
        CAF_L1CAM=c("L1CAM", "ITGB8", "CCBE1"),
        CAF_CXCL14=c("COL4A5", "CXCL14", "F3", "PDGFRA"),
        CAF_CTHRC1=c("CTHRC1", "FN1", "COL10A1", "FAP"),
        CAF_COL4=c("PDGFA", "COL4", "COL5A3", "LZTS1")
    )
)
```

::: {.panel-tabset}

```{r mgs-hm-sel, results="asis", fig.width=15, fig.height=4}
ps <- lapply(sub, \(.) {
    .gs <- unique(unlist(gs[[.]]))
    .gs <- intersect(.gs, rownames(lys[[.]]))
    mtx <- logcounts(lys[[.]])[.gs, ]
    .plt_hm(lys[[.]], mtx, by="kid", id=.) + 
        coord_flip(expand=FALSE) + theme(
            axis.text.y=element_text(size=12), 
            axis.text.x=element_text(size=8, angle=90, hjust=1, vjust=0.5))
})
for (. in sub) { cat("#### ", ., "\n"); print(ps[[.]]); cat("\n\n") }
```

:::

## downstreams

### signatures

```{r sig-lys}
# collect list of signatures from MSigDB
sig <- list(YAP="WP_HIPPOYAP_SIGNALING")
# hallmark gene sets (H)
ids <- c(
    apoptosis="APOPTOSIS",
    angiogenesis="ANGIOGENESIS",
    EMT="EPITHELIAL_MESENCHYMAL_TRANSITION",
    FA_metabolism="FATTY_ACID_METABOLISM",
    glycolysis="GLYCOLYSIS",
    hypoxia="HYPOXIA",
    infl_resp="INFLAMMATORY_RESPONSE",
    IFNa="INTERFERON_ALPHA_RESPONSE",
    IFNg="INTERFERON_GAMMA_RESPONSE",
    myogenesis="MYOGENESIS",
    STAT3="IL6_JAK_STAT3_SIGNALING",
    TNFa="TNFA_SIGNALING_VIA_NFKB",
    WNT="WNT_BETA_CATENIN_SIGNALING")
for (. in names(ids)) sig[[.]] <- paste0("HALLMARK_", ids[[.]])
# Hallmarks of transcriptional intratumour 
# heterogeneity across a thousand tumours
# Gavish et al., 2023 (DOI: 10.1038/s41586-023-06130-4)
ids <- c(
    IG="36_IG",
    MYC="20_MYC",
    stress="5_STRESS",
    respiration="21_RESPIRATION",
    epi_senescence="19_EPITHELIAL_SENESCENCE")
pre <- "GAVISH_3CA_MALIGNANT_METAPROGRAM_"
for (. in names(ids)) sig[[.]] <- paste0(pre, ids[.])
pre <- "GAVISH_3CA_METAPROGRAM_FIBROBLASTS_CAF_"
for (. in seq_len(10)) sig[[paste0("CAF_", .)]] <- paste0(pre, .)
# A human cell atlas of fetal gene expression
# Cao et al., 2020 (DOI: 10.1126/science.aba7721)
ids <- c(
    epithelial="INTESTINAL_EPITHELIAL_CELLS",
    endo.lympathic="LYMPHATIC_ENDOTHELIAL_CELLS",
    endo.vascular="VASCULAR_ENDOTHELIAL_CELLS",
    lymphoid="LYMPHOID_CELLS",
    mesothelial="MESOTHELIAL_CELLS",
    myeloid="MYELOID_CELLS",
    smooth.muscle="SMOOTH_MUSCLE_CELLS",
    stromal="STROMAL_CELLS")
pre <- "DESCARTES_FETAL_INTESTINE_"
for (. in names(ids)) sig[[paste0("fetal_", .)]] <- paste0(pre, ids[.])
length(sig)
```

```{r sig-get}
set <- lapply(sig, \(.) {
    tf <- tempfile(fileext=".json")
    url <- sprintf("https://www.gsea-msigdb.org/gsea/msigdb/human/download_geneset.jsp?geneSetName=%s&fileType=json", .)
    download.file(url, destfile=tf, mode="wb")
    gs <- jsonlite::fromJSON(tf)[[1]]$geneSymbols
    intersect(gs, rownames(sce))
})
sapply(set, length)
```

```{r sig-run}
auc <- lapply(sub, \(.) {
    mtx <- logcounts(lys[[.]])
    rnk <- AUCell_buildRankings(mtx, BPPARAM=bp, plotStats=FALSE, verbose=FALSE)
    AUCell_calcAUC(set, rnk, nCores=th, verbose=FALSE)
})
```

```{r sig-plt, fig.width=12, fig.height=6}
ps <- lapply(sub, \(.) .plt_hm(lys[[.]], auc[[.]], by=c("kid", "pid"), id=.))
wrap_plots(ps, nrow=1) + plot_layout(guides="collect")
```

### selected

```{r fig.width=12, fig.height=6}
mgs <- list(
    epi=list(
        endocrine=c("CHGA", "SST"),
        tuft=c("DCLK1", "TSLP", "CDHR2", "ESPN"),
        paneth=c("LYZ", "DEFA5", "DEFA6", "DLL4"),
        entero=c("ALPI", "FABP1", "SLC26A2", "CA1"),
        krt=grep("^KRT", rownames(sce), value=TRUE),
        neuron=c("PHOX2B", "CHRNA3", "HAND2", "TBX3"),
        goblet=c("MUC2", "TFF3", "SATB2", "KLK1", "SPINK4"),
        stem=c("OLFM4", "LGR5", "KCNE3", "KCNQ1", "PROM1", "ASCL2", "AXIN2"),
        other=c("REG1A", "REG1B", "REG3A", "REG3G", "REG4", "CEACAM5", "CEACAM6")
    ),
    imm=list(
        mast=c("KIT", "IL1RL1"),
        DC=c("LAMP3", "CLEC10A", "ITGAX"),
        mono=c("VCAN"),
        macro=c("CD14", "TLR4", "CD68", "APOE", "APOC1", "CTSD", "MMP9"),
        Th=c("CD3D", "CD2", "CD28", "CD4", "ITGAL", "CD40LG", "TIGIT", "GATA3"),
        Tc=c("CD8A", "NKG7", "EOMES", "GZMK", "CXCR6", "CCL5"),
        PC=c("IGKC", "IGHA1", "IGHA2", "IGHG1", "IGHG3", "IGHM", "MZB1", "XBP1", "IRF4"),
        BC=c("CD79A", "CD40", "CIITA", "CD83", "CD19", "MS4A1", "CD69", "SELL", "IGHD")
    ),
    str=list(
        SMC=c("MYL9", "MYLK", "MYH11", "CCN1"),
        EC=c("PECAM1", "CAV1", "VWF", "CD34", "CD93", "EGFL7", "KLF2"),
        BEC=c("CDH5", "FLT1", "CLDN5"),
        LEC=c("LYVE1", "PROX1", "CCL21", "PDPN"),
        fib=c("COL1A1", "COL6A1", "COL3A1", "COL6A2", "COL6A3",
            "ADAMDEC1", "FN1", "F3", "C3", "MMP1", "MMP3", "C1S", "CLU"),
        CAF=c("COL1A2", "FAP", "PDGFRA", "PDGFB", "MMP2", "DCN", "ACTA2", "TAGLN", "PDGFA")
    )
)
nan <- setdiff(unlist(mgs), rownames(sce))
if (length(nan)) {print(nan); stop()}
lapply(sub, \(.) {
    df <- data.frame(
        g=unlist(mgs[[.]]),
        k=rep.int(names(mgs[[.]]), sapply(mgs[[.]], length)))
    mtx <- logcounts(lys[[.]])[unlist(mgs[[.]]), ]
    p <- .plt_hm(lys[[.]], mtx, by=c("kid", "pid"), id=.) + 
        ggtitle(.lab(., length(unlist(mgs[[.]])))) +
        labs(fill="z-scaled\nexpr.") 
    q <- ggplot(df, aes(0, g, fill=k)) + 
        geom_tile() + labs(fill=.) + coord_equal(expand=FALSE) +
        theme_void() + theme(legend.key.size=unit(0.5, "lines")) +
        scale_fill_manual(values=hcl.colors(length(unique(df$k)), "Spectral"))
    (p | q) + 
        plot_layout(guides="collect") &
        scale_y_discrete(limits=p$scales$scales[[2]]$limits)
}) |> wrap_plots(nrow=1,) + plot_layout(nrow=1)
```

## appendix

### saving

```{r save-data}
for (. in sub) {
    rds <- sprintf("../outs/sub-%s.rds", .)
    if (!file.exists(rds)) saveRDS(lys[[.]], rds)
}
```

::: {.callout-note icon=false, collapse=true}

### session

```{r session-info}
#| code-fold: false
sessionInfo()
```

:::
